Initiate the yaml generation

This function acts as an entry point for the initialisation of the yaml generation.

init(
  authors = "author",
  date = lubridate::today(),
  affiliation = NULL,
  dir = ".",
  left = "yes",
  right = NULL,
  navbar_title = "Title",
  title = "siteymlgen",
  categories = NULL,
  name = NULL,
  output_dir = NULL,
  include = NULL,
  exclude = NULL,
  toc = FALSE,
  toc_depth = NULL,
  toc_title = NULL,
  toc_float = FALSE,
  theme = "united",
  highlight = NULL,
  css = NULL,
  file = NULL
)

Arguments

authors
A character or vector of characters of authors to add to the _site.yml

date
A character or a vector of characters describing the date

affiliation
A character or a vector of characters describing the authors affiliation. The affiliation which must match the length of the vector supplied to name.

dir
The directory of the Rmarkdown documents that you want automatically built into a _site.yml file.

left
A TRUE or FALSE value that determines if you want your navigation bar left aligned. This is the default location.

right
A TRUE or FALSE value that determines if you want your navigation bar right aligned. Left aligned is the default location

navbar_title
The title of the navigation bar. If none supplied this defaults to "Title"

title
A character describing the title of the website. This defaults to "siteymlgen" if no value is supplied

categories
A character or vector of characters describing the categories for the website.

name
A character or a vector of characters describing the name of the author(s)

output_dir
A character describing the output directory name for the final Rmarkdown website.

include
A character or vector of characters detailing the extra files to include in the final Rmarkdown build

exclude
A character or vector of characters detailing the files not to include in the final Rmarkdown build

toc
Specify whether a table of contents is required for your website, defaults to TRUE

toc_depth
An integer to specify the depth of the table of contents

toc_title
A character that sets the toable of contents title

toc_float
A boolean to specify if the table of contents is static or floats. defaults to FALSE.

theme
An Rmarkdown theme to set the render output. Defaults to 'united'.

highlight
Syntax highlighting style for html_output from the rmarkdown package. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", "haddock", and "textmate". Pass NULL to prevent syntax highlighting

css
Specify custom css.

file
Specify the output filename.

Value

A _site.yml file populated with yaml code.

Examples

init(authors = c("Adam","Paul"), left="yes",
navbar_title = "Main title", title = "hello",
categories=c("r", "reprodicibility"), file=tempdir())

#> Warning: cannot create file '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp5G961R', reason 'Is a directory'

#> Warning: No file variable has been specified. Please specify one.
#>       Typically init(file='_site.yml')